fix: change spriteId type from size_t to unsigned int in ShootEvent struct

+++ server/src/ecs/events/Events.hpp
@@ -41,7 +41,7 @@ struct ShootEvent {
     size_t shooter;                 ///> ID of the entity that fired the projectile
     std::pair<float, float> bounds; ///> Width and height of the projectile
     float lifetime;                 ///> Lifetime of the projectile in seconds
-    size_t spriteId;                ///> Sprite ID for rendering the projectile
+    unsigned int spriteId;          ///> Sprite ID for rendering the projectile
 };
 
 /**
